feature with adapter from db

+++ app/src/main/java/com/example/ken/updish/BackgroundWorker/PostDetailsBackgroundWorker.java
@@ -162,8 +162,8 @@ public class PostDetailsBackgroundWorker extends AsyncTask<String, Void, String>
             loc.setCity(obj.getString("city"));
             loc.setProvince(obj.getString("province"));
             // Ken
-            loc.setLatitude(obj.getLong("latitude"));
-            loc.setLongtitude(obj.getLong("longtitude"));
+            loc.setLatitude(obj.getDouble("latitude"));
+            loc.setLongtitude(obj.getDouble("longtitude"));
 
             tempPost.setLocation(loc);
